Update to syncer, upsert attachments

+++ database/data-resync.py
@@ -9,6 +9,8 @@ import os
 import botocore 
 import botocore.session 
 from aws_secretsmanager_caching import SecretCache, SecretCacheConfig
+import boto3
+import io
 
 def lambda_handler(event, context):
   try:
@@ -34,6 +36,8 @@ def lambda_handler(event, context):
     bucket_name = 'wfnews-dev-uploads'
     aws_key = ''
     aws_secret = ''
+    # Flush all data first?
+    data_flush = False
 
     print('Fetching a token from OAUTH...')
     token_response = requests.get(token_service, auth=HTTPBasicAuth(client_name, client_secret))
@@ -80,17 +84,18 @@ def lambda_handler(event, context):
 
       del wfim_response
 
-    # Flush the existing cache
-    print('Flushing the WFNEWS AWS cache')
-    wfnews_del_response = requests.delete(wfnews_api + 'publishedIncident/flush', headers={'Authorization': 'Bearer ' + token})
-    if wfnews_del_response.status_code != 204:
-      print(wfnews_del_response)
-      print('... Failed to flush')
-      return {
-          'statusCode': wfnews_del_response.status_code,
-          'body': json.dumps('Failed to fetch data from WFIM. Response code was: ' + str(wfnews_del_response.status_code))
-      }
-    del wfnews_del_response
+    # Flush the existing cache, if desired
+    if data_flush:
+      print('Flushing the WFNEWS AWS cache')
+      wfnews_del_response = requests.delete(wfnews_api + 'publishedIncident/flush', headers={'Authorization': 'Bearer ' + token})
+      if wfnews_del_response.status_code != 204:
+        print(wfnews_del_response)
+        print('... Failed to flush')
+        return {
+            'statusCode': wfnews_del_response.status_code,
+            'body': json.dumps('Failed to fetch data from WFIM. Response code was: ' + str(wfnews_del_response.status_code))
+        }
+      del wfnews_del_response
 
     print('... Create Inserts for WFNEWS')
     # iterate incidents, push to WFNEWS API with some random data
@@ -99,17 +104,16 @@ def lambda_handler(event, context):
       published_incident = None
 
       # check for existing published data and fetch that
-      wfim_response = requests.get(wfim_api + '/publishedIncidents/byIncident/' + incident['wildfireIncidentGuid'],  headers={'Authorization': 'Bearer ' + token})
-      if wfim_response.status_code != 200 and wfim_response.status_code != 404:
+      wfim_response = requests.get(wfim_api + 'publishedIncidents/byIncident/' + incident['wildfireIncidentGuid'],  headers={'Authorization': 'Bearer ' + token})
+      if wfim_response.status_code != 200:
         print('Failed to load published info. Response code was: ' + str(wfim_response.status_code))
-      elif wfim_response.status_code == 200:
+      else:
         published_incident = wfim_response.json()
-        print(published_incident)
 
       del wfim_response
 
       feature = {
-        "publishedIncidentDetailGuid": str(uuid.uuid4()),
+        "publishedIncidentDetailGuid": published_incident['publishedIncidentDetailGuid'] if published_incident is not None else incident['wildfireIncidentGuid'], #str(uuid.uuid4()),
         "incidentGuid": incident['wildfireIncidentGuid'],
         "incidentNumberLabel": incident['incidentNumberSequence'], # number sequence or incidentLabel?
         "newsCreatedTimestamp": curr_time,
@@ -127,22 +131,22 @@ def lambda_handler(event, context):
         "newsPublicationStatusCode": published_incident['newsPublicationStatusCode'] if published_incident is not None else 'DRAFT',
         "incidentOverview": published_incident['incidentOverview'] if published_incident is not None else '',
         "traditionalTerritoryDetail": published_incident['traditionalTerritoryDetail'] if published_incident is not None else None,
-        "incidentSizeType": published_incident['incidentSizeType'] if published_incident is not None else '',
+        "incidentSizeType": '',
         "incidentSizeDetail": published_incident['incidentSizeDetail'] if published_incident is not None else None,
         "incidentCauseDetail": published_incident['incidentCauseDetail'] if published_incident is not None else None,
         "contactOrgUnitIdentifer": published_incident['contactOrgUnitIdentifer'] if published_incident is not None else '',
         "contactPhoneNumber": published_incident['contactPhoneNumber'] if published_incident is not None else '',
         "contactEmailAddress": published_incident['contactEmailAddress'] if published_incident is not None else '',
         "resourceDetail": published_incident['resourceDetail'] if published_incident is not None else None,
-        "wildfireCrewResourcesInd": True if published_incident is not None and published_incident['wildfireCrewResourcesInd'] == 'true' else False,
+        "wildfireCrewResourcesInd": True if published_incident is not None and published_incident['wildfireCrewResourcesInd'] == True else False,
         "wildfireCrewResourcesDetail": published_incident['wildfireCrewResourcesDetail'] if published_incident is not None else None,
-        "wildfireAviationResourceInd": True if published_incident is not None and published_incident['wildfireAviationResourceInd'] == 'true' else False,
+        "wildfireAviationResourceInd": True if published_incident is not None and published_incident['wildfireAviationResourceInd'] == True else False,
         "wildfireAviationResourceDetail": published_incident['wildfireAviationResourceDetail'] if published_incident is not None else None,
-        "heavyEquipmentResourcesInd": True if published_incident is not None and published_incident['heavyEquipmentResourcesInd'] == 'true' else False,
+        "heavyEquipmentResourcesInd": True if published_incident is not None and published_incident['heavyEquipmentResourcesInd'] == True else False,
         "heavyEquipmentResourcesDetail": published_incident['heavyEquipmentResourcesDetail'] if published_incident is not None else None,
-        "incidentMgmtCrewRsrcInd": True if published_incident is not None and published_incident['incidentMgmtCrewRsrcInd'] == 'true' else False,
+        "incidentMgmtCrewRsrcInd": True if published_incident is not None and published_incident['incidentMgmtCrewRsrcInd'] == True else False,
         "incidentMgmtCrewRsrcDetail": published_incident['incidentMgmtCrewRsrcDetail'] if published_incident is not None else None,
-        "structureProtectionRsrcInd": True if published_incident is not None and published_incident['structureProtectionRsrcInd'] == 'true' else False,
+        "structureProtectionRsrcInd": True if published_incident is not None and published_incident['structureProtectionRsrcInd'] == True else False,
         "structureProtectionRsrcDetail": published_incident['structureProtectionRsrcDetail'] if published_incident is not None else None,
         "publishedTimestamp": published_incident['publishedTimestamp'] if published_incident is not None else curr_time,
         "publishedUserTypeCode": published_incident['publishedUserTypeCode'] if published_incident is not None else 'GOV', 
@@ -155,6 +159,7 @@ def lambda_handler(event, context):
         "fireYear": fire_year
       }
 
+      # The API will update if the posted resource exists
       wfnews_response = requests.post(wfnews_api + 'publishedIncident', json=feature, headers={'Authorization': 'Bearer ' + token, 'content-type': 'application/json'})
       # verify 200
       if wfnews_response.status_code != 201:
@@ -191,31 +196,71 @@ def lambda_handler(event, context):
       else:
         attachments = wfim_response.json()
         for attachment in attachments['collection']:
-          print(attachment)
           attachment['@type'] = 'AttachmentResource'
           attachment['imageURL'] = '/' + str(incident['incidentNumberSequence']) + '/' + attachment['fileName']
           attachment['attachmentTypeCode'] = 'DOCUMENT'
           attachment['sourceObjectNameCode'] = 'PUB_INC_DT'
+          attachment['sourceObjectUniqueId'] = incident['incidentNumberSequence']
+          # apply the mime type correctly - not coming accross
 
           wfnews_response = requests.post(wfnews_api + 'publishedIncidentAttachment/' + str(incident['incidentNumberSequence']) + '/attachments', json=attachment, headers={'Authorization': 'Bearer ' + token, 'content-type': 'application/json'})
           if wfnews_response.status_code != 201:
             print('Failed to upload attachment')
           else:
             print('... Attachment ref created. Fetching from WFDM and writing to bucket...')
+            new_attachment = wfnews_response.json()
             # Fetch attachment from WFDM
             wfdm_response = requests.get(wfdm_api + 'documents/' + str(attachment['fileIdentifier']) + '/bytes', headers={'Authorization': 'Bearer ' + token })
             doc_bytes = wfdm_response.content
-            # write bytes to s3 bucket
-            client = boto3.client('s3',
-              region_name='ca-central-1',
-              aws_access_key_id=aws_key,
-              aws_secret_access_key=aws_secret)
-            client.put_object(Body=doc_bytes, Bucket=bucket_name, Key=attachment['imageURL'])
-
             del wfdm_response
+            wfnews_response = requests.post(wfnews_api + 'publishedIncidentAttachment/' + str(incident['incidentNumberSequence']) + '/attachments/' + new_attachment['attachmentGuid'] + '/bytes', files=dict(file = io.BytesIO(doc_bytes)), headers={'Authorization': 'Bearer ' + token})
           del wfnews_response
 
       del wfim_response
+
+    # Once the Inserts and updates are complete, see if anything needs to be removed
+    # note, this is only needed if we did not flush data
+    if not data_flush:
+      # load the data from the public api
+      wfnews_incidents = []
+      page_number = 0
+      total_pages = 1
+      page_size = 100
+      exit_while = False
+      while page_number < total_pages and exit_while == False:
+        page_number = page_number + 1
+        wfnews_response = requests.get(wfnews_api + 'publicPublishedIncident?&pageNumber=' + str(page_number)+ '&pageRowCount=' + str(page_size))
+        if wfnews_response.status_code != 200:
+          return {
+            'statusCode': wfnews_response.status_code,
+            'body': json.dumps('Failed to fetch data from WFNEWS. Response code was: ' + str(wfnews_response.status_code))
+          } 
+
+        cache_data = wfnews_response.json()
+
+        total_pages = cache_data['totalPageCount']
+        wfnews_incidents = wfnews_incidents + cache_data['collection']
+
+        if cache_data['pageRowCount'] == 0 or len(cache_data['collection']) < page_size:
+          exit_while = True
+
+        del wfnews_response
+
+      if len(wfnews_incidents) != len(wfim_incidents):
+        # different lengths, so we need to delete things
+        print('... Deleting unmatched incidents')
+        for incident in wfnews_incidents:
+          found_match = False
+          for wfim_incident in wfim_incidents:
+            if wfim_incident['wildfireIncidentGuid'] == incident['incidentGuid']:
+              found_match = True
+              break
+          if not found_match:
+            #incident doesn't have a match, so we can delete it
+            wfnews_response = requests.delete(wfnews_api + 'publishedIncident/' + incident['publishedIncidentDetailGuid'], headers={'Authorization': 'Bearer ' + token})
+            if wfnews_response.status_code != 200:
+              print('Failed to delete')
+            del wfnews_response
 
     return {
       'statusCode': 200,
